+++ src/Google/AdsApi/AdWords/v201708/cm/AdGroupBidModifierService.php
@@ -1,0 +1,121 @@
+<?php
+
+namespace Google\AdsApi\AdWords\v201708\cm;
+
+
+/**
+ * This file was generated from WSDL. DO NOT EDIT.
+ */
+class AdGroupBidModifierService extends \Google\AdsApi\Common\AdsSoapClient
+{
+
+    /**
+     * @var array $classmap The defined classes
+     */
+    private static $classmap = array (
+      'AuthenticationError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\AuthenticationError',
+      'AuthorizationError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\AuthorizationError',
+      'ClientTermsError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\ClientTermsError',
+      'CriterionError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\CriterionError',
+      'DateRange' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\DateRange',
+      'DistinctError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\DistinctError',
+      'EntityNotFound' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\EntityNotFound',
+      'FieldPathElement' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\FieldPathElement',
+      'IdError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\IdError',
+      'InternalApiError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\InternalApiError',
+      'NewEntityCreationError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\NewEntityCreationError',
+      'NotEmptyError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\NotEmptyError',
+      'OperationAccessDenied' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\OperationAccessDenied',
+      'OperatorError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\OperatorError',
+      'OrderBy' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\OrderBy',
+      'Paging' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Paging',
+      'Platform' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Platform',
+      'Predicate' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Predicate',
+      'PreferredContent' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\PreferredContent',
+      'QueryError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\QueryError',
+      'QuotaCheckError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\QuotaCheckError',
+      'RangeError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\RangeError',
+      'RateExceededError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\RateExceededError',
+      'ReadOnlyError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\ReadOnlyError',
+      'RejectedError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\RejectedError',
+      'RequestError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\RequestError',
+      'RequiredError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\RequiredError',
+      'SelectorError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\SelectorError',
+      'SizeLimitError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\SizeLimitError',
+      'SoapHeader' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\SoapHeader',
+      'SoapResponseHeader' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\SoapResponseHeader',
+      'StringFormatError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\StringFormatError',
+      'StringLengthError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\StringLengthError',
+      'DatabaseError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\DatabaseError',
+      'ApiError' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\ApiError',
+      'ApiException' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\ApiException',
+      'ApplicationException' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\ApplicationException',
+      'Selector' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Selector',
+      'Criterion' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Criterion',
+      'AdGroupBidModifier' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\AdGroupBidModifier',
+      'AdGroupBidModifierOperation' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\AdGroupBidModifierOperation',
+      'AdGroupBidModifierPage' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\AdGroupBidModifierPage',
+      'AdGroupBidModifierReturnValue' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\AdGroupBidModifierReturnValue',
+      'ListReturnValue' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\ListReturnValue',
+      'Operation' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Operation',
+      'Page' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\Page',
+      'getResponse' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\getResponse',
+      'mutateResponse' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\mutateResponse',
+      'queryResponse' => 'Google\\AdsApi\\AdWords\\v201708\\cm\\queryResponse',
+    );
+
+    /**
+     * @param array $options A array of config values
+     * @param string $wsdl The wsdl file to use
+     */
+    public function __construct(array $options = array(),
+                $wsdl = 'https://adwords.google.com/api/adwords/cm/v201708/AdGroupBidModifierService?wsdl')
+    {
+      foreach (self::$classmap as $key => $value) {
+        if (!isset($options['classmap'][$key])) {
+          $options['classmap'][$key] = $value;
+        }
+      }
+      $options = array_merge(array (
+      'features' => 1,
+    ), $options);
+      parent::__construct($wsdl, $options);
+    }
+
+    /**
+     * Gets ad group level criterion bid modifiers.
+     *
+     * @param \Google\AdsApi\AdWords\v201708\cm\Selector $selector
+     * @return \Google\AdsApi\AdWords\v201708\cm\AdGroupBidModifierPage
+     * @throws \Google\AdsApi\AdWords\v201708\cm\ApiException
+     */
+    public function get(\Google\AdsApi\AdWords\v201708\cm\Selector $selector)
+    {
+      return $this->__soapCall('get', array(array('selector' => $selector)))->getRval();
+    }
+
+    /**
+     * Adds, removes or updates ad group bid modifier overrides.
+     *
+     * @param \Google\AdsApi\AdWords\v201708\cm\AdGroupBidModifierOperation[] $operations
+     * @return \Google\AdsApi\AdWords\v201708\cm\AdGroupBidModifierReturnValue
+     * @throws \Google\AdsApi\AdWords\v201708\cm\ApiException
+     */
+    public function mutate(array $operations)
+    {
+      return $this->__soapCall('mutate', array(array('operations' => $operations)))->getRval();
+    }
+
+    /**
+     * Returns a list of {@link AdGroupBidModifier}s that match the query.
+     *
+     * @param string $query
+     * @return \Google\AdsApi\AdWords\v201708\cm\AdGroupBidModifierPage
+     * @throws \Google\AdsApi\AdWords\v201708\cm\ApiException
+     */
+    public function query($query)
+    {
+      return $this->__soapCall('query', array(array('query' => $query)))->getRval();
+    }
+
+}