+++ src/Google/AdsApi/AdWords/v201708/cm/PolicySummaryReviewState.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Google\AdsApi\AdWords\v201708\cm;
+
+
+/**
+ * This file was generated from WSDL. DO NOT EDIT.
+ */
+class PolicySummaryReviewState
+{
+    const UNKNOWN = 'UNKNOWN';
+    const REVIEW_IN_PROGRESS = 'REVIEW_IN_PROGRESS';
+    const REVIEWED = 'REVIEWED';
+
+
+}