+++ src/Google/AdsApi/AdWords/v201708/cm/PolicyTopicEvidenceType.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Google\AdsApi\AdWords\v201708\cm;
+
+
+/**
+ * This file was generated from WSDL. DO NOT EDIT.
+ */
+class PolicyTopicEvidenceType
+{
+    const UNKNOWN = 'UNKNOWN';
+    const AD_TEXT = 'AD_TEXT';
+    const HTTP_CODE = 'HTTP_CODE';
+    const WEBSITES = 'WEBSITES';
+    const LANGUAGE = 'LANGUAGE';
+
+
+}