+++ src/Google/AdsApi/AdWords/v201708/o/IdeaType.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Google\AdsApi\AdWords\v201708\o;
+
+
+/**
+ * This file was generated from WSDL. DO NOT EDIT.
+ */
+class IdeaType
+{
+    const KEYWORD = 'KEYWORD';
+
+
+}