+++ src/Google/AdsApi/AdWords/v201708/o/RequestType.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Google\AdsApi\AdWords\v201708\o;
+
+
+/**
+ * This file was generated from WSDL. DO NOT EDIT.
+ */
+class RequestType
+{
+    const IDEAS = 'IDEAS';
+    const STATS = 'STATS';
+
+
+}